Fixed bugs in OpenGL 2.1 mode launch

+++ boar-main/src/main/kotlin/net/spartanb312/everett/graphics/drawing/RenderUtils.kt
@@ -1008,7 +1008,7 @@ object RenderUtils : Renderer2D by if (RS.compatMode) ArrayedRenderer2D else PMV
         lineWidth: Float = 1f,
         color: ColorRGB,
     ) = drawLinesStrip0(
-        getArcVertices(centerX.toFloat(), centerY.toFloat(), radius, angleRange, segments),
+        getArcVertices(centerX, centerY, radius, angleRange, segments),
         lineWidth,
         color
     )
